Make bindings extensible
