+++ ampersand-dom-bindings.js
@@ -1,58 +1,6 @@
 /*$AMPERSAND_VERSION*/
-var Store = require('key-tree-store');
-var dom = require('ampersand-dom');
-var matchesSelector = require('matches-selector');
-var partial = require('lodash/partial');
-var slice = Array.prototype.slice;
-
-function getMatches(el, selector, firstOnly) {
-    if (selector === '') return [el];
-    var matches = [];
-    if (!selector) return matches;
-    if (firstOnly) {
-        if (matchesSelector(el, selector)) return [el];
-        return el.querySelector(selector) ? [el.querySelector(selector)] : [];
-    } else {
-        if (matchesSelector(el, selector)) matches.push(el);
-        return matches.concat(slice.call(el.querySelectorAll(selector)));
-    }
-}
-function setAttributes(el, attrs) {
-    for (var name in attrs) {
-        dom.setAttribute(el, name, attrs[name]);
-    }
-}
-
-function removeAttributes(el, attrs) {
-    for (var name in attrs) {
-        dom.removeAttribute(el, name);
-    }
-}
-
-function makeArray(val) {
-    return Array.isArray(val) ? val : [val];
-}
-
-function switchHandler(binding, el, value) {
-    // the element selector to show
-    var showValue = binding.cases[value];
-
-    var firstMatchOnly = binding.firstMatchOnly;
-
-    // hide all the other elements with a different value
-    for (var item in binding.cases) {
-        var curValue = binding.cases[item];
-
-        if (value !== item && curValue !== showValue) {
-            getMatches(el, curValue, firstMatchOnly).forEach(function (match) {
-                dom.hide(match);
-            });
-        }
-    }
-    getMatches(el, showValue, firstMatchOnly).forEach(function (match) {
-        dom.show(match);
-    });
-}
+var BindingStore = require('./bindings-store');
+var KeyStore = require('key-tree-store');
 
 function getSelector(binding) {
     if (typeof binding.selector === 'string') {
@@ -64,182 +12,27 @@ function getSelector(binding) {
     }
 }
 
+// Get the binding name to access from the binding store
+function getType(type) {
+    if (type) {
+        if (typeof type === 'string') {
+            return type;
+        } else if (typeof type === 'function') {
+            return 'function';
+        }
+    }
+
+    //returning the passed type by default maintains backwards compatability for throwing 'no such binding type: '
+    return type;
+}
+
+// Returns a curried version of the binding store function to update the dom when coressponding events are fired
 function getBindingFunc(binding, context) {
-    var type = binding.type || 'text';
-    var isCustomBinding = typeof type === 'function';
+    var type = getType(binding.type);
     var selector = getSelector(binding);
-    var firstMatchOnly = binding.firstMatchOnly;
-    var yes = binding.yes;
-    var no = binding.no;
-    var hasYesNo = !!(yes || no);
-
-    // storage variable for previous if relevant
-    var previousValue;
-
-    if (isCustomBinding) {
-        return function (el, value) {
-            getMatches(el, selector, firstMatchOnly).forEach(function (match) {
-                type.call(context, match, value, previousValue);
-            });
-            previousValue = value;
-        };
-    } else if (type === 'text') {
-        return function (el, value) {
-            getMatches(el, selector, firstMatchOnly).forEach(function (match) {
-                dom.text(match, value);
-            });
-        };
-    } else if (type === 'class') {
-        return function (el, value) {
-            getMatches(el, selector, firstMatchOnly).forEach(function (match) {
-                dom.switchClass(match, previousValue, value);
-            });
-            previousValue = value;
-        };
-    } else if (type === 'attribute') {
-        if (!binding.name) throw Error('attribute bindings must have a "name"');
-        return function (el, value) {
-            var names = makeArray(binding.name);
-            getMatches(el, selector, firstMatchOnly).forEach(function (match) {
-                names.forEach(function (name) {
-                    dom.setAttribute(match, name, value);
-                });
-            });
-            previousValue = value;
-        };
-    } else if (type === 'value') {
-        return function (el, value) {
-            getMatches(el, selector, firstMatchOnly).forEach(function (match) {
-                if (!value && value !== 0) value = '';
-                // only apply bindings if element is not currently focused
-                if (document.activeElement !== match) match.value = value;
-            });
-            previousValue = value;
-        };
-    } else if (type === 'booleanClass') {
-        // if there's a `no` case this is actually a switch
-        if (hasYesNo) {
-            yes = makeArray(yes || '');
-            no = makeArray(no || '');
-            return function (el, value) {
-                var prevClass = value ? no : yes;
-                var newClass = value ? yes : no;
-                getMatches(el, selector, firstMatchOnly).forEach(function (match) {
-                    prevClass.forEach(function (pc) {
-                        dom.removeClass(match, pc);
-                    });
-                    newClass.forEach(function (nc) {
-                        dom.addClass(match, nc);
-                    });
-                });
-            };
-        } else {
-            return function (el, value, keyName) {
-                var name = makeArray(binding.name || keyName);
-                var invert = (binding.invert || false);
-                value = (invert ? (value ? false : true) : value);
-                getMatches(el, selector, firstMatchOnly).forEach(function (match) {
-                    name.forEach(function (className) {
-                        dom[value ? 'addClass' : 'removeClass'](match, className);
-                    });
-                });
-            };
-        }
-    } else if (type === 'booleanAttribute') {
-        // if there are `yes` and `no` selectors, this swaps between them
-        if (hasYesNo) {
-            yes = makeArray(yes || '');
-            no = makeArray(no || '');
-            return function (el, value) {
-                var prevAttribute = value ? no : yes;
-                var newAttribute = value ? yes : no;
-                getMatches(el, selector, firstMatchOnly).forEach(function (match) {
-                    prevAttribute.forEach(function (pa) {
-                        if (pa) {
-                            dom.removeAttribute(match, pa);
-                        }
-                    });
-                    newAttribute.forEach(function (na) {
-                        if (na) {
-                            dom.addAttribute(match, na);
-                        }
-                    });
-                });
-            };
-        } else {
-            return function (el, value, keyName) {
-                var name = makeArray(binding.name || keyName);
-                var invert = (binding.invert || false);
-                value = (invert ? (value ? false : true) : value);
-                getMatches(el, selector, firstMatchOnly).forEach(function (match) {
-                    name.forEach(function (attr) {
-                        dom[value ? 'addAttribute' : 'removeAttribute'](match, attr);
-                    });
-                });
-            };
-        }
-    } else if (type === 'toggle') {
-        var mode = (binding.mode || 'display');
-        var invert = (binding.invert || false);
-        // this doesn't require a selector since we can pass yes/no selectors
-        if (hasYesNo) {
-            return function (el, value) {
-                getMatches(el, yes, firstMatchOnly).forEach(function (match) {
-                    dom[value ? 'show' : 'hide'](match, mode);
-                });
-                getMatches(el, no, firstMatchOnly).forEach(function (match) {
-                    dom[value ? 'hide' : 'show'](match, mode);
-                });
-            };
-        } else {
-            return function (el, value) {
-                value = (invert ? (value ? false : true) : value);
-                getMatches(el, selector, firstMatchOnly).forEach(function (match) {
-                    dom[value ? 'show' : 'hide'](match, mode);
-                });
-            };
-        }
-    } else if (type === 'switch') {
-        if (!binding.cases) throw Error('switch bindings must have "cases"');
-        return partial(switchHandler, binding);
-    } else if (type === 'innerHTML') {
-        return function (el, value) {
-            getMatches(el, selector, firstMatchOnly).forEach(function (match) {
-                dom.html(match, value);
-            });
-        };
-    } else if (type === 'switchClass') {
-        if (!binding.cases) throw Error('switchClass bindings must have "cases"');
-        return function (el, value, keyName) {
-            var name = makeArray(binding.name || keyName);
-            for (var item in binding.cases) {
-                getMatches(el, binding.cases[item], firstMatchOnly).forEach(function (match) {
-                    name.forEach(function (className) {
-                        dom[value === item ? 'addClass' : 'removeClass'](match, className);
-                    });
-                });
-            }
-        };
-    } else if (type === 'switchAttribute') {
-        if (!binding.cases) throw Error('switchAttribute bindings must have "cases"');
-        return function (el, value, keyName) {
-            getMatches(el, selector, firstMatchOnly).forEach(function (match) {
-                if (previousValue) {
-                    removeAttributes(match, previousValue);
-                }
-
-                if (value in binding.cases) {
-                    var attrs = binding.cases[value];
-                    if (typeof attrs === 'string') {
-                        attrs = {};
-                        attrs[binding.name || keyName] = binding.cases[value];
-                    }
-                    setAttributes(match, attrs);
-
-                    previousValue = attrs;
-                }
-            });
-        };
+    
+    if (BindingStore.hasKey(type)) {
+        return BindingStore.get(type).call(context, binding, selector);
     } else {
         throw new Error('no such binding type: ' + type);
     }
@@ -251,7 +44,7 @@ function getBindingFunc(binding, context) {
 // all resulting functions should be called
 // like func(el, value, lastKeyName)
 module.exports = function (bindings, context) {
-    var store = new Store();
+    var store = new KeyStore();
     var key, current;
 
     for (key in bindings) {
